add endpoint to test paged list serialization in .net 8
