+++ Paging.sln
@@ -14,7 +14,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging.Benchmarks", "benchm
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging.ConsoleTests", "tests\Paging.ConsoleTests\Paging.ConsoleTests.csproj", "{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging.WebApplicationTest", "tests\Paging.WebApplicationTest\Paging.WebApplicationTest.csproj", "{F0777CC1-F809-4A46-8085-D418EA2DD14A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Paging.WebApplicationTest.Net8", "tests\Paging.WebApplicationTest.Net8\Paging.WebApplicationTest.Net8.csproj", "{F0777CC1-F809-4A46-8085-D418EA2DD14A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApplication1", "tests\WebApplication1\WebApplication1.csproj", "{5599C753-6E1F-4698-96A3-076CDB749A38}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,7 @@ Global
 		{D797B4BE-4C73-42A2-908C-208CC1244967} = {D7066DB3-706B-4D98-A0B1-D2553BCB61A2}
 		{EE5EB4CB-4B2F-49B6-AFF9-3066D2F335EC} = {DCA445CD-81D1-461E-8066-923FB17477EB}
 		{F0777CC1-F809-4A46-8085-D418EA2DD14A} = {DCA445CD-81D1-461E-8066-923FB17477EB}
+		{5599C753-6E1F-4698-96A3-076CDB749A38} = {DCA445CD-81D1-461E-8066-923FB17477EB}
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{C85597B6-A4C7-4E0C-854B-66FD5F27F33B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -49,5 +52,9 @@ Global
 		{F0777CC1-F809-4A46-8085-D418EA2DD14A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F0777CC1-F809-4A46-8085-D418EA2DD14A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F0777CC1-F809-4A46-8085-D418EA2DD14A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5599C753-6E1F-4698-96A3-076CDB749A38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5599C753-6E1F-4698-96A3-076CDB749A38}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5599C753-6E1F-4698-96A3-076CDB749A38}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5599C753-6E1F-4698-96A3-076CDB749A38}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
