+++ tests/Paging.WebApplicationTest.Net8/Paging.WebApplicationTest.Net8.csproj
@@ -5,11 +5,16 @@
 		<Nullable>enable</Nullable>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<InvariantGlobalization>true</InvariantGlobalization>
+		<LangVersion>default</LangVersion>
 	</PropertyGroup>
 
 	<ItemGroup>
 		<PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.0"/>
 		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0"/>
+	</ItemGroup>
+ 
+	<ItemGroup>
+	  <ProjectReference Include="..\..\src\Paging\Paging.csproj" />
 	</ItemGroup>
 
 </Project>
