+++ tests/Paging.WebApplicationTest.Net8/Paging.WebApplicationTest.Net8.http
@@ -1,6 +1,7 @@
 @Paging.WebApplicationTest_HostAddress = http://localhost:5084
 
-GET {{Paging.WebApplicationTest_HostAddress}}/weatherforecast/
+GET {{Paging.WebApplicationTest_HostAddress}}/paging?pageNumber=1&pageSize=10
 Accept: application/json
+
 
 ###
