+++ tests/Paging.WebApplicationTest.Net8/Program.cs
@@ -1,0 +1,36 @@
+using Paging.Extensions;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+	app.UseSwagger();
+	app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+app.MapGet(
+		"/paging",
+		(int pageNumber, int pageSize) =>
+		{
+			var list = new List<int> {
+				1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24,
+				25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50
+			};
+			
+			return list.Paginated(pageNumber, pageSize);
+		}
+	)
+	.WithName("PagedList")
+	.WithOpenApi();
+
+app.Run();
